fix: move GeoPeerTable encode method to parser module for consistency
- Move encode() implementation from models to parser module
- Consistent with PeerIndexTable and RibAfiEntries pattern
- Remove feature gate as bytes is always available in parser
- Move encoding test to parser module
- Fixes build with --no-default-features
