+++ src/parser/mrt/messages/table_dump_v2/geo_peer_table.rs
@@ -3,7 +3,7 @@
 use crate::error::ParserError;
 use crate::models::*;
 use crate::parser::ReadUtils;
-use bytes::{Buf, Bytes};
+use bytes::{Buf, BufMut, Bytes, BytesMut};
 use std::net::IpAddr;
 
 /// Parse GEO_PEER_TABLE message according to RFC 6397
@@ -107,6 +107,85 @@ pub fn parse_geo_peer_table(data: &mut Bytes) -> Result<GeoPeerTable, ParserErro
     }
 
     Ok(geo_table)
+}
+
+impl GeoPeerTable {
+    /// Encode the GEO_PEER_TABLE into bytes according to RFC 6397
+    ///
+    /// # Returns
+    ///
+    /// A `Bytes` object containing the encoded GEO_PEER_TABLE data.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use std::net::Ipv4Addr;
+    /// use std::str::FromStr;
+    /// use bgpkit_parser::models::{GeoPeerTable, GeoPeer, Peer, Asn};
+    ///
+    /// let mut geo_table = GeoPeerTable::new(
+    ///     Ipv4Addr::from_str("10.0.0.1").unwrap(),
+    ///     "test-view".to_string(),
+    ///     51.5074,  // London latitude
+    ///     -0.1278,  // London longitude
+    /// );
+    ///
+    /// let encoded = geo_table.encode();
+    /// ```
+    pub fn encode(&self) -> Bytes {
+        let mut buf = BytesMut::new();
+
+        // Encode collector BGP ID (4 bytes)
+        buf.put_u32(self.collector_bgp_id.into());
+
+        // Encode view name length and view name
+        let view_name_bytes = self.view_name.as_bytes();
+        buf.put_u16(view_name_bytes.len() as u16);
+        buf.extend(view_name_bytes);
+
+        // Encode collector coordinates (4 bytes each, 32-bit float)
+        buf.put_f32(self.collector_latitude);
+        buf.put_f32(self.collector_longitude);
+
+        // Encode peer count
+        buf.put_u16(self.geo_peers.len() as u16);
+
+        // Encode each peer entry
+        for geo_peer in &self.geo_peers {
+            // Encode peer type (1 byte)
+            buf.put_u8(geo_peer.peer.peer_type.bits());
+
+            // Encode peer BGP ID (4 bytes)
+            buf.put_u32(geo_peer.peer.peer_bgp_id.into());
+
+            // Encode peer IP address (4 or 16 bytes depending on address family)
+            match geo_peer.peer.peer_ip {
+                std::net::IpAddr::V4(ipv4) => {
+                    buf.put_u32(ipv4.into());
+                }
+                std::net::IpAddr::V6(ipv6) => {
+                    buf.extend_from_slice(&ipv6.octets());
+                }
+            }
+
+            // Encode peer AS number (2 or 4 bytes depending on AS size)
+            if geo_peer
+                .peer
+                .peer_type
+                .contains(crate::models::PeerType::AS_SIZE_32BIT)
+            {
+                buf.put_u32(u32::from(geo_peer.peer.peer_asn));
+            } else {
+                buf.put_u16(u16::from(geo_peer.peer.peer_asn));
+            }
+
+            // Encode peer coordinates (4 bytes each, 32-bit float)
+            buf.put_f32(geo_peer.peer_latitude);
+            buf.put_f32(geo_peer.peer_longitude);
+        }
+
+        buf.freeze()
+    }
 }
 
 #[cfg(test)]
@@ -309,5 +388,80 @@ mod tests {
 
         // Test overall equality
         assert_eq!(parsed_table, original_table);
+    }
+
+    #[test]
+    fn test_geo_peer_table_encoding() {
+        let collector_bgp_id = Ipv4Addr::from_str("10.0.0.1").unwrap();
+        let mut geo_table = GeoPeerTable::new(
+            collector_bgp_id,
+            "test-view".to_string(),
+            51.5074, // London latitude
+            -0.1278, // London longitude
+        );
+
+        // Add a peer with IPv4 address and 2-byte AS
+        let peer1 = Peer::new(
+            Ipv4Addr::from_str("1.1.1.1").unwrap(),
+            Ipv4Addr::from_str("2.2.2.2").unwrap().into(),
+            Asn::new_16bit(65001),
+        );
+        let geo_peer1 = GeoPeer::new(peer1, 40.7128, -74.0060); // New York
+        geo_table.add_geo_peer(geo_peer1);
+
+        // Add a peer with IPv6 address and 4-byte AS
+        let peer2 = Peer::new(
+            Ipv4Addr::from_str("3.3.3.3").unwrap(),
+            std::net::Ipv6Addr::from_str("2001:db8::1").unwrap().into(),
+            Asn::new_32bit(65002),
+        );
+        let geo_peer2 = GeoPeer::new(peer2, f32::NAN, f32::NAN); // Private coordinates
+        geo_table.add_geo_peer(geo_peer2);
+
+        // Encode the geo table
+        let encoded = geo_table.encode();
+
+        // Create expected bytes manually for comparison
+        let mut expected = BytesMut::new();
+
+        // Collector BGP ID
+        expected.put_u32(0x0A000001); // 10.0.0.1
+
+        // View name length and name
+        let view_name = "test-view";
+        expected.put_u16(view_name.len() as u16);
+        expected.extend_from_slice(view_name.as_bytes());
+
+        // Collector coordinates
+        expected.put_f32(51.5074);
+        expected.put_f32(-0.1278);
+
+        // Peer count
+        expected.put_u16(2);
+
+        // First peer: IPv4, 2-byte AS
+        expected.put_u8(0x00); // Peer type: IPv4, 2-byte AS
+        expected.put_u32(0x01010101); // BGP ID: 1.1.1.1
+        expected.put_u32(0x02020202); // Peer IP: 2.2.2.2
+        expected.put_u16(65001); // AS number
+        expected.put_f32(40.7128); // New York latitude
+        expected.put_f32(-74.0060); // New York longitude
+
+        // Second peer: IPv6, 4-byte AS
+        expected.put_u8(0x03); // Peer type: IPv6, 4-byte AS
+        expected.put_u32(0x03030303); // BGP ID: 3.3.3.3
+        expected.extend_from_slice(&[
+            0x20, 0x01, 0x0d, 0xb8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x00, 0x01,
+        ]); // IPv6: 2001:db8::1
+        expected.put_u32(65002); // AS number
+        expected.put_f32(f32::NAN); // Private latitude
+        expected.put_f32(f32::NAN); // Private longitude
+
+        let expected_bytes = expected.freeze();
+
+        // Compare the encoded bytes with expected
+        assert_eq!(encoded.len(), expected_bytes.len());
+        assert_eq!(encoded, expected_bytes);
     }
 }
